Implemented GCounter
Implemented GCounter based on the implementation described by Shapiro et al. (2011) in the paper "A comprehensive study of Convergent and Commutative Replicated Data Types". Uses the objects VersionVector as the state of the counter.

SummingAccumulators which were required for the GCounter implementation have been provided,

OpenIMAJ references have been implemented to reference the original implementation described in literature.