Replaced SummingAccumulator with Arithmetic
Replaced the SummingAccumulator interface and implementations with a more generic Arithmetic interface
allowing for a more broad range of arithmetic operations to be added. Surrently this has been expanded with
a sub method allowing for subtraction. The sum method has been replaced with an add method of the same
functionality.

An ArithmeticAbstractTest has been created and the two implementations IntegerArithmetic and LongArithmetic
both take advantage of this with the IntegerArithmeticTest and LongArithmeticTest.