Implemented CommutativeGSet
CmRDT interface requires the update message to be a VersionedUpdateMessage.

Created an AbstractCmRDT providing the delivery preconditions for all recieved updates as well as synching the local version after a successful update.

Tests have been created for ensuring the commutativity of the implementations (currently supporting grow-only sets) and common functionality from the ConvergentSetAbstractTest has been abstracted out along with some refactoring of the name of the convergence tests.