Refactored growable set tests
Common code from Convergent and Commutative Updatable Set Tests have been moved into UpdatableSetAbstractTest with the addition of several abstract methods to provide the objects and assertions that would be specific to certain implementations. This includes replacing the enums which were used for a similar purpose with methods which is advantageous as they can be created as abstract if implementation is required. There is no longer a getZeroVerion method with an alternative method for creating the VersionVector for testing to be implemented.

GrowOnlySet_ classes have now been renamed to GrowableSet_ to repsect the fact that non-grow-only implementations extend this: the new name is no longer contradictary in that scenario.

Similar changes for remove based operations are yet to be implemented.