Test refactoring and duplication removal
Refactored tests such that multiple tests classes are used for each class being tested. This has allowed a reduction in code duplication which can be fruther extended and will ensure that all new tests written affect all of the classes which they are relevant to.

Enums have been completely removed and replaced with abstract or overridable methods.