Implemented Dots
A Dot interface has been provided and tested with methods allowing the ability to get Dots from VersionVectors and to perform simple comparisons between the two.

The isDotted methods have been removed from the VersionVector interface and its implementations along with any dotted variables and constructor arguments. This change has been refactored throughout the codebase.

A mutable flag has been added to the VersionVectorAbstractTest to allow the tests to adapt to the type being tested.

Added tests for successor method. Added the type in the toString method of AbstractVersion so that they can be identified.