Implemented Dots for Commutative implementations
Added the required methods which were missing from the Dot/VersionVector implementations for identical and precedes.

Versions have had additional generics to define the type that can be compared against and the type which clones are. This has now removed the ability to compare a VersionVector with another version containing a Map and provides stronger guarantees with the symetry of methods such as identical. As part of this, a LamportTimestamp interface has been created for pure implementations of LogicalTimestamp. For objects which have a Version associated with them, a Versioned interface has been created.

Refactoring has beed done throughout the codebase to support the new generics and interfaces that were added.

DottedUpdateMessage and AbtractDottedUpdateMessage have also been added which are used for UpdateMessages utilising Dots as the version. These are then used by the commutative implementations to ensure that the UpdateMessages have dots and that the semantics related to dots are used when performing operations. With Dots being used as part of the commutative implementations, all tests pass.