Removed redundant generics/method
Redudant generics were being included which were not required and provided no use other than making code look messy and making it harder to maintain. This was primarily version types which were used within other generic types. Since the version type is not used directly, knowing the type provided no benefit.

Additionally getDot() was a redudant method. All it did was provide a type checked version but since the Version is given as one of the generic parameters, getVersion() returns the correctly typed version without the extra layer of indirection.