CmRDT can use any Version
Previously a constraint was imposed that required CmRDT to use Dots (as this was the only type that correctly implemented precedence). To accomodate this change, AbstractCmRDT has been made more general with AbstractDottedCmRDT taking the role AbstractCmRDT previously held providing the abstract base for objects which use dots for versioning.

Some minor optimisations have been made in AbstactCmRDT which assumes that calls to the update are likely to be valid and recuding the number of checks performed before being able to apply that change. If the message is a duplicate or out of order, then 1 or 2 additional checks are performed respectively.

Documentation for CmRDT has been extensivly improved explaining in more detail what a CmRDT is, what the requirements for delivery are, etc. Other minor documentation tweaks have been applied to delivery components.