Delivery now has StateDeliveryChannel and ReliableDeliveryChannel
The DeliveryChannel implementations have been split into two types: StateDeliveryChannel and ReliableDeliveryChannel. ReliableDeliveryChannel is similar to the pre-existing implementations with StateDeliveryChannel allowing for weaker conditions.

The DeliveryChannel interface is still the base interface for these two different types and an AbstractDeliveryChannel contains all the commonality between the two. In order to isolate responsibility and allow delivery implementations to write as little code as possible and still implement both types, DeliveryExchange has been created.

DeliveryExchange provideds reliable delivery between DeliveryChannels and is used by both of the DeliveryChannel types. DeliveryChannels are no longer an encapsulation of the underlying delivery method but of delivery semantics and the guarantees provided. The implementation provided for each type, PeriodicStateDeliveryChnnel and PeriodicReliableDeliveryChannel can use any compatible exchange and each provide the semantics required for the datatypes they support.

These implementations both use a scheduled runnable which runs at a fixed rate in order to perform their background processing. Because of the design, it would be possible to implement alternative delivery methods.

In addition to the periodic implementations, Null implementations have been provided which do not exchange messages. This are useful for testing and being able to develop applications without having to setup the DeliveryExchange.

StateDeliveryChannels work by being notified of changes and then taking a snapshot of the Updtable to deliver to other replicas whenever they wish. This provides a huge performance improvement and significantly reduces memory usage compared to creating a snapshot for every mutation. Snapshotting is an intensive process and should be limited.

The new types have been applied to the datatypes with CvRDTs requiring a StateDeliveryChannel and CmRDTs requiring a ReliableDeliveryChannel. Abstraction has been used to assist with the integration of these DeliveryChannels.

Tests have been created for the new interfaces. With a much smaller scope of repsonsibility for the DeliveryChannel thanks to the introduction of the DeliveryExchnage. Unit tests which test just the functionality and semantics of the DeliveryChannel and DeliveryExchnage ensure each component works idependantly before being integrated together.

Other tests have been updated to use the correct types of DeliveryChannel when testing and these pass successfully.

The previously used LocalDeliveryChannel was single threaded in implementation. With the new implementations, everything has been implemented asynchrously which eliminated the deadlock condition which existed when using the LocalDeliveryChannel in a multi-threaded environment.

To assist with testing, the wait functionality of the DeliveryUtils which is possible through the implementation of the DeliveryChannel/DeliveryExchange have been enhance. There is now a guranteed overlapping of the wait functions between the source channel, the intermediatry exchange(s) and the destination of the messages allowing for a reliable method for waiting on delivery. For the most part, this is implemented with locks on the objects but there is still one part which requires busy sleeping. This can be improved upon in the future.